solve: [PGM] 67257 - 수식최대화  풀이
